// src/models/medicineDP.interface.ts
export interface IMedicineDP {
  medicineSubContract?: string | any;
  medicineId?: string | any;
  materialId?: string | any;
  medicineName?: string | any;
  description?: string | any;
  location?: string | any;
  quantity?: number | any;
  shipper?: string | any;
  manufacturer?: string | any;
  distributor?: string | any;
  pharma?: string | any;
  packageStatus?: string | number | any;
  transactionBlocksDP?: any;
  saleStatus?: string | number | any;
}